+++ junit-jupiter-api/src/main/java/module-info.java
@@ -30,6 +30,7 @@ module org.junit.jupiter.api {
 	exports org.junit.jupiter.api.function;
 	exports org.junit.jupiter.api.io;
 	exports org.junit.jupiter.api.parallel;
+	exports org.junit.jupiter.api.util to org.junit.jupiter.engine;
 
 	opens org.junit.jupiter.api.condition to org.junit.platform.commons;
 }
