+++ junit-jupiter-api/src/main/java/org/junit/jupiter/api/util/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Internal JUnit Jupiter utilities.
+ *
+ * <h2>DISCLAIMER</h2>
+ *
+ * <p>These utilities are intended solely for usage within the JUnit framework
+ * itself. <strong>Any usage by external parties is not supported.</strong>
+ * Use at your own risk!
+ */
+
+@NullMarked
+package org.junit.jupiter.api.util;
+
+import org.jspecify.annotations.NullMarked;
