+++ platform-tooling-support-tests/projects/jar-describe-module/junit-jupiter-api.expected.txt
@@ -12,4 +12,5 @@ requires org.apiguardian.api static transitive
 requires org.jspecify static transitive
 requires org.junit.platform.commons transitive
 requires org.opentest4j transitive
+qualified exports org.junit.jupiter.api.util to org.junit.jupiter.engine
 qualified opens org.junit.jupiter.api.condition to org.junit.platform.commons
